requirements/app.txt
Django==1.10.4
djangorestframework==3.5.3
django-extensions==1.7.5
elizabeth==0.3.9
gunicorn==19.6.0
psycopg2==2.6.2
django-rest-swagger==2.1.0
